requirements/requirements.txt
Django==4.1.13
django-filter==23.5
django-jazzmin==2.6.2
django-rest-knox==4.2.0
djangorestframework==3.15.1
djongo==1.3.6
drf-yasg==1.21.7
Faker==26.0.0
gunicorn[gevent]==22.0.0
pymongo[srv]==3.12.3
pytz==2024.1
ssm-parameter-store==19.11.0
Unipath==1.1
